Moved code from Data/Graph to Securities/CurrencyGraph
